app/requirements.txt
Flask==0.12.2
pygithub==1.38